Moving oof flags to dedicated flags.rs
